Time: 0 ms (100.00%) | Memory: 10.4 MB (73.91%) - LeetSync

+++ 410-split-array-largest-sum/split-array-largest-sum.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     // HELPER: Checks if we can split 'nums' into 'k' or fewer subarrays 
-    // such that no single subarray sum exceeds 'mid'.
+    // without any single subarray sum exceeding 'mid'.
     bool isPossible(vector<int>& nums, int k, int mid) {
         int split = 1; // Start with 1 subarray
         int sum = 0;   // Current subarray sum
 
         for (int i = 0; i < nums.size(); i++) {
-            // If adding this number fits within limit 'mid', pack it in
+            // If adding this number fits within limit 'mid', keep packing
             if (sum + nums[i] <= mid) {
                 sum += nums[i];
             } 
@@ -17,36 +17,38 @@ public:
                 sum = nums[i];
             }
 
-            if(split > k){
+            // EARLY EXIT: The moment splits exceed 'k', 'mid' is invalid.
+            // Stop checking the rest of the array immediately!
+            if (split > k) {
                 return false;
             }
         }
 
-        // True if we can fit everything into 'k' or fewer subarrays
-        return true;
+        return true; // Successfully split into <= k subarrays
     }
 
     int splitArray(vector<int>& nums, int k) {
-        // MINIMUM POSSIBLE ANSWER:
-        // A single subarray must hold at least the largest single number.
+        // MINIMUM BOUND:
+        // Max single element (a subarray must at least hold the largest number).
         int low = *max_element(nums.begin(), nums.end());
 
-        // MAXIMUM POSSIBLE ANSWER:
-        // If k = 1 (no splits), the answer is the total sum of all numbers.
-        int high = accumulate(nums.begin(), nums.end(), 0);
+        // MAXIMUM BOUND:
+        // Sum of all elements (worst-case scenario where k = 1).
+        // Note: Using 0LL prevents 32-bit integer overflow during summation.
+        long long high = accumulate(nums.begin(), nums.end(), 0);
 
         int ans = high;
 
-        // Binary Search on the ANSWER range [max_element ... total_sum]
+        // Binary search on the answer range [max_element ... total_sum]
         while (low <= high) {
             int mid = low + (high - low) / 2;
 
             // DECISION RULE:
             if (isPossible(nums, k, mid)) {
-                ans = mid;       // 'mid' works! Save it as a valid answer.
-                high = mid - 1;  // Try searching LEFT for an even smaller max sum.
+                ans = mid;       // 'mid' works! Save it as a candidate answer.
+                high = mid - 1;  // Search LEFT for an even smaller max sum.
             } else {
-                low = mid + 1;   // 'mid' was too small (needed > k splits). Search RIGHT.
+                low = mid + 1;   // 'mid' was too small (exceeded k splits). Search RIGHT.
             }
         }
